feat: Crear la ui de bienvenida, meter las imagenes de la aplicación, y modificar ciertos fichero .java que estaban mal hechos

+++ app/src/main/java/com/sthenos/fortium/model/entities/Serie.java
@@ -94,7 +94,7 @@ public class Serie {
         this.ejercicioId = ejercicioId;
     }
 
-    public int getPeso() {
+    public float getPeso() {
         return peso;
     }
 
@@ -110,7 +110,7 @@ public class Serie {
         this.repeticiones = repeticiones;
     }
 
-    public int getRpeRir() {
+    public float getRpeRir() {
         return rpeRir;
     }
